Create the add and list method for the music and genre class
