+++ class/music.rb
@@ -1,10 +1,11 @@
 require_relative 'items'
 
 class MusicAlbum < Item
- attr_accessor :on_spotify 
- attr_reader :public_date
+ attr_accessor :on_spotify, :public_date
+ attr_reader :id
  def initialize(on_spotify,public_date)
   super(public_date)
+  @public_date = public_date
   @on_spotify = on_spotify
   @items = []
  end
